pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "autowing"
version = "0.6.2"
description = "auto-wing is a tool that uses LLM to assist automated testing."
readme = "README.md"
authors = [
    { name = "defnngj", email = "defnngj@gmail.com" }
]
requires-python = ">=3.10,<4.0"

dependencies = [
    "openai>=2.15.0, <3.0.0",
    "dashscope>=1.25.9,<2.0.0",
    "google-genai (>=1.62.0,<2.0.0)",
    "python-dotenv>=1.2.1,<2.0.0",
    "loguru>=0.7.3,<0.8.0",
]

[project.urls]
repository = "https://github.com/SeldomQA/auto-wing"
homepage = "https://github.com/SeldomQA/auto-wing"
